CLJS-2866: Predicate-induced type inference
Look for if tests that look like the simple application of a predicate
to a local, and in that case, map certain predicates to the tags that
satisfying that predicate implies. For example a (string? x) test
implies that x has the tag string in the then branch of an if.

Likewise check for the use of satisfies? and instance? and induce
tags accordingly based on the type argument.

For core predicates that delegate to satisfies? or instance?,
we have hard-coded entries in the table that essentially cause things
to behave as if the predicate were inlined. For example, (counted? x)
behaves just like (satisfies? ICounted x).

We only induce tags if no tag previously exists (or a tag exists but
it is the special any tag.) This ensures that we don't override type
hints, or generally revise existing logic. (We really only induce a tag
as a last resort.)

Arrange the code so that we can add more complicated induced tags in
the future (perhaps allowing for logical connectives in the test, for
example.)