CLJS-2869: Improve type inference of and / or
Adds tighter type inference for the and / or macros. In general, the
set of types inferred should be a subset of the set of types inferred
by the previous algorithm, and should be the minimal subset consistent
with the types pass to and / or.

Adds a few utilities for managing type tags, and refactors if
inference to use a new add-types function.

We also need to associate dynamic vars (if not type hinted) as having
any type (as opposed to being the type of their init) because otherwise
the common pattern of (def ^:dynamic *foo* nil) resulting in a tag of
clj-nil can lead to inferrence of compound expressions involving such
dynamic vars. In a few places where the init is actually a boolean,
we add a type hint of boolean to preserve previous behavior.