CLJS-2869: Improve type inference of and / or
Adds tighter type inference for the and / or macros. This is
accomplished via improved type inference for the if special form, as
well as new type inference for the and / or macros in the case that
they don't expand to make use of if.

In general, the set of types inferred should be a subset of the set of
types inferred by the previous algorithm, and should be the minimal
subset consistent with the types pass to and / or / if (verified via
exhaustive tests that generate all values for types of interest,
comparing inference to actual behavior of and / or / if.)

We also need to associate dynamic vars (if not type hinted) as having
any type (as opposed to being the type of their init) because otherwise
the common pattern of (def ^:dynamic *foo* nil) resulting in a tag of
clj-nil can lead to inferrence of compound expressions involving such
dynamic vars. In a few places where the init is actually a boolean,
we add a type hint of boolean to preserve previous behavior.