CLJS-2883: Instrumentation fails compilation with a large number of spec'd functions
When instrumenting, unstrumenting, or checking, we normally evaluate
the suppied form in order to convert it to an (unquoted) symbol or
collection of symbols. In the case that this is all speced vars or all
checkable syms, this quoted list can be large enough to cause eval to
fail. Fortunately, in this special case there is no need to actually
apply eval because we know the list of symbols is the second item in
the form.