CLJS-2917: Failure to load user.cljs if temp out dir
I initially tried to solve the issue by fixing cljs.repl/load-file so it
would pass cljs.closure/compile-file a relative path string, but then I
saw that cljs.repl.rhino,nashorn,node,graaljs does the same thing, and
would presumably fail if also used with temp out dir. It seems like it's
expected that compile-file should be able to handle a File with absolute
path, so I modified it to do that.