CLJS-2994 Ensure all prepl :vals are pr-str-ed
The only time this should ever fire is if an exception is caught during
evaluation. All other ClojureScript evaluation results are already
strings, so we should only ever have to run results of Throwable->map
through pr-str.

Performing this in the valf allows this to be tweaked by users where
required. It also acts as a catch all to ensure _any_ future values that
happen to flow through to :val as maps etc instead of strings will get
converted.

:val should _always_ be a string. This ensures that assumption will
always hold true.