CLJS-3056: runtime namespace load order is independent from ordering in ns macro :require form
Ensures that the :requires construction in cljs.compiler/emit-source
maintains the ordering that namespaces were required in the code being
compiled.

Currently, emit-source throws the analyzed dependencies into a Clojure
set, which does not preserve ordering. The patch uses the same approach
as a similar fix for CLJS-1453.

This patch makes the assumption that the underlying dependencies being used
here are in order, so I believe that this is fix is necessary, but possibly not
sufficient.