CLJS-3173: case statement with embedded list test constant
Calling `const?` on an embedded list, `cljs.analyzer/resolve-var` is
called on the last element of the list which is why this is raised:

    java.lang.Long cannot be cast to clojure.lang.Named

An extra condition is added so that only vars are being sent to
`resolve-var`:

    (core/and ...
              (core/var? (last x))
              (ana/resolve-var env (last x)))

This leads to another error:

    TypeError: Cannot find function call in object <MY_INTEGER>. (<cljs repl>clojure#1)

Which comes from the macroexpansion creating something like

    (cond
        (= (1 2) (quote (3 4)) ...))

There's already a bit in the preprocessing of the test-then pairs where
anything that's a symbol is quoted. Just need to add `seq` as another
instance where that's needed.

    ...
    (core/let [test (if (core/or (core/symbol? test) (seq? test))
                      (core/list 'quote test)
                      test)]